create bot logic + tests

+++ chat_bot/fly_generator.py
@@ -1,10 +1,12 @@
 import json
+import os
 import random
 import datetime
 
 CITIES_FILE = "json/avia_yandex_cites_id.json"
 RESULT_FILE = "json/result.json"
-EXCLUDE_FLIGHT_TRACKS = ["москва-мельбурн", "санкт-петербург-мельбурн", "мельбурн-москва", "мельбурн-санкт-петербург"]
+EXCLUDE_FLIGHT_TRACKS = ["москва - мельбурн", "санкт-петербург - мельбурн", "мельбурн - москва",
+                         "мельбурн - санкт-петербург"]
 
 
 def get_json(json_file):
@@ -18,8 +20,16 @@ def get_json(json_file):
 
 def write_to_json(source_dict):
     """:param source_dict json json file with result dict"""
-    with open(RESULT_FILE, 'w', encoding='utf-8') as fp:
-        json.dump(source_dict, fp, indent=4, ensure_ascii=False)
+    if os.path.isfile(RESULT_FILE):
+        with open(RESULT_FILE, 'r', encoding='utf-8') as fp:
+            data = json.load(fp)
+            data.update(source_dict)
+        with open(RESULT_FILE, 'w', encoding='utf-8') as fp:
+            fp.seek(0)
+            json.dump(data, fp, indent=4, ensure_ascii=False)
+    else:
+        with open(RESULT_FILE, 'w', encoding='utf-8') as fp:
+            json.dump(source_dict, fp, indent=4, ensure_ascii=False)
 
 
 def generate_sch_flights(count, from_town, to_town, week_days=None, month_days=None):
@@ -29,9 +39,9 @@ def generate_sch_flights(count, from_town, to_town, week_days=None, month_days=N
     :param week_days - list of days of week for scheduled
     :param month_days - list of day of month for schedule
     :returns dict of scheduled flights"""
-    key = f"{from_town}-{to_town}"
+    key = f"{from_town} - {to_town}"
     result = {key: []}
-    time = f"{random.randrange(0, 24):02d}:{random.randrange(0, 60, 5):02d} "
+    time = f"{random.randrange(0, 24):02d}:{random.randrange(0, 60, 5):02d}"
     date = datetime.datetime.now().replace(day=1)
     while count > 0:
         if key in EXCLUDE_FLIGHT_TRACKS:
@@ -65,11 +75,11 @@ def generate_random_flights(count, c_list):
     result = {}
     while count > 0:
         from_city, to_city = random.sample(set(c_list), 2)
-        key = f"{from_city}-{to_city}"
+        key = f"{from_city} - {to_city}"
         if key in EXCLUDE_FLIGHT_TRACKS:
             break
-        time = f"{random.randrange(0, 24):02d}:{random.randrange(0, 60, 5):02d} "
-        additional_val = f"2020-{random.randrange(1, 12):02d}-{random.randrange(1, 30):02d} {time}"
+        time = f"{random.randrange(0, 24):02d}:{random.randrange(0, 60, 5):02d}"
+        additional_val = f"{random.randrange(1, 30):02d}-{random.randrange(1, 12):02d}-2020 {time}"
         if key in result.keys():
             val = result.get(key)
             val.append(additional_val)
@@ -101,8 +111,8 @@ def generate_flight_dict(cites_list, weekly_scheduled_flight_count,
 yandex_dict = get_json(CITIES_FILE)
 cities_list = list(map(lambda x: x.lower(), [*yandex_dict]))
 # lists [[count,[week day\ month day, ...]],...]
-w_sch_fl_count = [[10, [1, 4]], [12, [2, 6]], [30, [1]], [5, [1, 3, 6]]]
-m_sch_fl_count = [[5, [5, 15, 25]], [4, [10, 20]], [6, [7]]]
+w_sch_fl_count = [[100, [1, 4]], [122, [2, 6]], [30, [1]], [52, [1, 3, 6]]]
+m_sch_fl_count = [[55, [5, 15, 25]], [44, [10, 20]], [56, [7]]]
 dict_to_json = generate_flight_dict(cites_list=cities_list, weekly_scheduled_flight_count=w_sch_fl_count,
-                                    monthly_scheduled_flight_count=m_sch_fl_count, random_flight_count=100)
+                                    monthly_scheduled_flight_count=m_sch_fl_count, random_flight_count=1000)
 write_to_json(dict_to_json)
